First commit for new cleanly organized R6RS code
This is the first commit for the new, cleaned and organized collection
of R6RS-modules (so far these are to be run under DrRacket). Just one
batch of actual Sequence-code now included, in
Seqs/Triangles/triangles-core.ss
All the *.scm files for now on are old modules intended for MIT/GNU
Scheme.
The R6RS-modules have now prefix *.ss